refactor: share wallet log component
